init new structure
