+++ src/Routing/Routes.tsx
@@ -1,0 +1,8 @@
+import React, { FC, useContext } from "react";
+import { BrowserRouter as Router, Switch, Route, Link } from "react-router-dom";
+import { LoginPage } from "../Pages/LoginPage";
+import { GamePage } from "../Pages/GamePage";
+
+const Routes: FC = () => {
+  return <div></div>;
+};
